+++ examples/rust/src/bin/client.rs
@@ -256,6 +256,10 @@ struct ActionSubscribe {
     #[clap(long)]
     slots_filter_by_commitment: bool,
 
+    /// Subscribe on interslot slot updates
+    #[clap(long)]
+    slots_interslot_updates: bool,
+
     /// Subscribe on transactions updates
     #[clap(long)]
     transactions: bool,
@@ -452,6 +456,7 @@ impl Action {
                         "client".to_owned(),
                         SubscribeRequestFilterSlots {
                             filter_by_commitment: Some(args.slots_filter_by_commitment),
+                            interslot_updates: Some(args.slots_interslot_updates),
                         },
                     );
                 }
@@ -699,7 +704,7 @@ async fn geyser_subscribe(
                         Some(UpdateOneof::Ping(_)) => (&mut pb_pp_c, &pb_pp),
                         Some(UpdateOneof::Pong(_)) => (&mut pb_pp_c, &pb_pp),
                         None => {
-                            error!("update not found in the message");
+                            pb_multi.println("update not found in the message")?;
                             break;
                         }
                     };
@@ -742,8 +747,9 @@ async fn geyser_subscribe(
                             let dir = "grpc-client-verify";
                             let name = SystemTime::now().duration_since(UNIX_EPOCH)?.as_nanos();
                             let path = format!("{dir}/{name}");
-                            error!("found unmached message, save to `{path}`");
-                            fs::create_dir(dir)
+                            pb_multi
+                                .println(format!("found unmached message, save to `{path}`"))?;
+                            fs::create_dir_all(dir)
                                 .await
                                 .context("failed to create dir for unmached")?;
                             fs::write(path, encoded_prost)
@@ -781,7 +787,8 @@ async fn geyser_subscribe(
                             json!({
                                 "slot": msg.slot,
                                 "parent": msg.parent,
-                                "status": status.as_str_name()
+                                "status": status.as_str_name(),
+                                "deadError": msg.dead_error,
                             }),
                         );
                     }