+++ yellowstone-grpc-proto/src/plugin/filter/message.rs
@@ -1,10 +1,10 @@
 use {
     crate::{
         geyser::{
-            subscribe_update::UpdateOneof, CommitmentLevel as CommitmentLevelProto,
-            SubscribeUpdate, SubscribeUpdateAccount, SubscribeUpdateAccountInfo,
-            SubscribeUpdateBlock, SubscribeUpdateEntry, SubscribeUpdatePing, SubscribeUpdatePong,
-            SubscribeUpdateSlot, SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
+            subscribe_update::UpdateOneof, SlotStatus as SlotStatusProto, SubscribeUpdate,
+            SubscribeUpdateAccount, SubscribeUpdateAccountInfo, SubscribeUpdateBlock,
+            SubscribeUpdateEntry, SubscribeUpdatePing, SubscribeUpdatePong, SubscribeUpdateSlot,
+            SubscribeUpdateTransaction, SubscribeUpdateTransactionInfo,
             SubscribeUpdateTransactionStatus,
         },
         plugin::{
@@ -27,7 +27,12 @@ use {
     prost_types::Timestamp,
     smallvec::SmallVec,
     solana_sdk::signature::Signature,
-    std::{collections::HashSet, sync::Arc, time::SystemTime},
+    std::{
+        collections::HashSet,
+        ops::{Deref, DerefMut},
+        sync::Arc,
+        time::SystemTime,
+    },
 };
 
 #[inline]
@@ -170,9 +175,10 @@ impl FilteredUpdate {
                 is_startup: msg.is_startup,
             }),
             FilteredUpdateOneof::Slot(msg) => UpdateOneof::Slot(SubscribeUpdateSlot {
-                slot: msg.0.slot,
-                parent: msg.0.parent,
-                status: msg.0.status as i32,
+                slot: msg.slot,
+                parent: msg.parent,
+                status: msg.status as i32,
+                dead_error: msg.dead_error.clone(),
             }),
             FilteredUpdateOneof::Transaction(msg) => {
                 UpdateOneof::Transaction(SubscribeUpdateTransaction {
@@ -543,31 +549,52 @@ impl FilteredUpdateAccount {
 #[derive(Debug, Clone, PartialEq)]
 pub struct FilteredUpdateSlot(MessageSlot);
 
+impl Deref for FilteredUpdateSlot {
+    type Target = MessageSlot;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl DerefMut for FilteredUpdateSlot {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
 impl prost::Message for FilteredUpdateSlot {
     fn encode_raw(&self, buf: &mut impl BufMut) {
-        let status = CommitmentLevelProto::from(self.0.status) as i32;
-        if self.0.slot != 0u64 {
-            ::prost::encoding::uint64::encode(1u32, &self.0.slot, buf);
+        let status = SlotStatusProto::from(self.status) as i32;
+        if self.slot != 0u64 {
+            ::prost::encoding::uint64::encode(1u32, &self.slot, buf);
         }
-        if let ::core::option::Option::Some(ref value) = self.0.parent {
+        if let ::core::option::Option::Some(ref value) = self.parent {
             ::prost::encoding::uint64::encode(2u32, value, buf);
         }
-        if status != CommitmentLevelProto::default() as i32 {
+        if status != SlotStatusProto::default() as i32 {
             ::prost::encoding::int32::encode(3u32, &status, buf);
+        }
+        if let Some(error) = &self.dead_error {
+            ::prost::encoding::string::encode(4u32, error, buf);
         }
     }
 
     fn encoded_len(&self) -> usize {
-        let status = CommitmentLevelProto::from(self.0.status) as i32;
+        let status = SlotStatusProto::from(self.status) as i32;
 
-        (if self.0.slot != 0u64 {
-            ::prost::encoding::uint64::encoded_len(1u32, &self.0.slot)
+        (if self.slot != 0u64 {
+            ::prost::encoding::uint64::encoded_len(1u32, &self.slot)
         } else {
             0
-        }) + self.0.parent.as_ref().map_or(0, |value| {
+        }) + self.parent.as_ref().map_or(0, |value| {
             ::prost::encoding::uint64::encoded_len(2u32, value)
-        }) + if status != CommitmentLevelProto::default() as i32 {
+        }) + if status != SlotStatusProto::default() as i32 {
             ::prost::encoding::int32::encoded_len(3u32, &status)
+        } else {
+            0
+        } + if let Some(error) = &self.dead_error {
+            ::prost::encoding::string::encoded_len(4u32, error)
         } else {
             0
         }
@@ -959,8 +986,8 @@ pub mod tests {
             plugin::{
                 filter::{name::FilterName, FilterAccountsDataSlice},
                 message::{
-                    CommitmentLevel, MessageAccount, MessageAccountInfo, MessageBlockMeta,
-                    MessageEntry, MessageSlot, MessageTransaction, MessageTransactionInfo,
+                    MessageAccount, MessageAccountInfo, MessageBlockMeta, MessageEntry,
+                    MessageSlot, MessageTransaction, MessageTransactionInfo, SlotStatus,
                 },
             },
         },
@@ -1238,9 +1265,13 @@ pub mod tests {
         for slot in [0, 42] {
             for parent in [None, Some(0), Some(42)] {
                 for status in [
-                    CommitmentLevel::Processed,
-                    CommitmentLevel::Confirmed,
-                    CommitmentLevel::Finalized,
+                    SlotStatus::Processed,
+                    SlotStatus::Confirmed,
+                    SlotStatus::Finalized,
+                    SlotStatus::FirstShredReceived,
+                    SlotStatus::Completed,
+                    SlotStatus::CreatedBank,
+                    SlotStatus::Dead,
                 ] {
                     encode_decode_cmp(
                         &["123"],
@@ -1248,10 +1279,21 @@ pub mod tests {
                             slot,
                             parent,
                             status,
+                            dead_error: None,
                             created_at: Timestamp::from(SystemTime::now()),
                         }),
                     )
                 }
+                encode_decode_cmp(
+                    &["123"],
+                    FilteredUpdateOneof::slot(MessageSlot {
+                        slot,
+                        parent,
+                        status: SlotStatus::Dead,
+                        dead_error: Some("123".to_owned()),
+                        created_at: Timestamp::from(SystemTime::now()),
+                    }),
+                )
             }
         }
     }